feat: blog post #3 + edited 'build.sh' script to select which page build

+++ build.sh
@@ -4,6 +4,19 @@
 dir=src # $dir = `.\src` folder
 dest=public # $dest = `.\public` folder
 
+if [ -z "$1" ]; then
+  PAGES="ALL"
+else
+  if [ "$1" = "BLOG" ]; then
+    PAGES="BLOG"
+  elif [ "$1" = "LEETCODE" ]; then
+    PAGES="LEETCODE"
+  else
+    PAGES="ALL"
+  fi
+fi
+
+
 # create a directory at $dest, 
 # -p = means if any of the parent directories does not exist, then `mkdir` will create them without raising an error
 mkdir -p $dest
@@ -11,15 +24,19 @@ mkdir -p $dest
 # copy (-R = recursively) all the files from $dir 
 cp -R $dir/* $dest
 # remove (-rf = forcefully and recursively) these folders
-rm -rf $dest/blog
 rm -rf $dest/parts
-rm -rf $dest/leetcode
 
-# build thoughts
-$dir/blog/build.sh
+if [[ $PAGES = "BLOG" || $PAGES = "ALL" ]]; then
+  # build thoughts
+  rm -rf $dest/blog
+  $dir/blog/build.sh
+fi
 
-# build leetcode
-$dir/leetcode/build.sh
+if [[ $PAGES = "LEETCODE" || $PAGES = "ALL" ]]; then
+  # build leetcode
+  rm -rf $dest/leetcode
+  $dir/leetcode/build.sh
+fi
 
 # replace <!-- NAV --> with src/parts/nav.html
 echo "BUILDING: navbar"
